chore!: @callstack/licenses to support both browser & node targets, move around files
