+++ packages/license-kit/package.json
@@ -54,8 +54,10 @@
   },
   "devDependencies": {
     "@callstack/license-kit-visualizer": "workspace:^",
+    "@callstack/licenses": "workspace:^",
     "@types/express": "^5.0.3",
     "@types/jest": "^29.5.5",
+    "@types/signale": "^1.4.7",
     "jest": "^29.7.0",
     "rimraf": "^6.0.1",
     "tsx": "^4.20.3",
@@ -68,6 +70,7 @@
     "helmet": "^8.1.0",
     "next": "^15.3.4",
     "open": "^10.1.2",
+    "signale": "^1.4.0",
     "ts-markdown-builder": "^0.4.1"
   }
 }
