+++ packages/license-kit/src/commands/copyleft.ts
@@ -2,16 +2,14 @@ import fs from 'node:fs';
 import path from 'node:path';
 import process from 'node:process';
 
-import { scanDependencies } from '@callstack/licenses';
+import {
+  STRONG_COPYLEFT_LICENSES_LOWERCASE,
+  WEAK_COPYLEFT_LICENSES_LOWERCASE,
+  scanDependencies,
+} from '@callstack/licenses';
 import type { Command } from 'commander';
 
-import {
-  ERROR_EMOJI,
-  NON_TAB_HELP_LISTING_SUBLIST_OFFSET,
-  STRONG_COPYLEFT_LICENSES,
-  WARNING_EMOJI,
-  WEAK_COPYLEFT_LICENSES,
-} from '../constants';
+import { ERROR_EMOJI, NON_TAB_HELP_LISTING_SUBLIST_OFFSET, WARNING_EMOJI } from '../constants';
 import { createScanOptionsFactory } from '../scanOptionsUtils';
 import { curryCommonScanOptions, validateCommonScanOptions } from '../utils/commandUtils';
 
@@ -49,11 +47,11 @@ export default function copyleftCommandSetup(program: Command): Command {
         continue;
       }
 
-      if (STRONG_COPYLEFT_LICENSES.has(value.type)) {
+      if (STRONG_COPYLEFT_LICENSES_LOWERCASE.has(value.type.toLowerCase())) {
         strongCopyleftLicensesFound.push(`- ${value.name}: ${value.type} (${value.file || value.url})`);
       }
 
-      if (WEAK_COPYLEFT_LICENSES.has(value.type)) {
+      if (WEAK_COPYLEFT_LICENSES_LOWERCASE.has(value.type.toLowerCase())) {
         weakCopyleftLicensesFound.push(`- ${value.name}: ${value.type} (${value.file || value.url})`);
       }
     }
