+++ packages/license-kit/visualizer/src/constants.ts
@@ -1,5 +1,3 @@
-import { LicenseCategory } from './types/LicenseCategory';
-
 export const SIDEBAR_EXPANDED_WIDTH = 500;
 export const SIDEBAR_COLLAPSED_WIDTH = 50;
 
@@ -13,12 +11,5 @@ export const NODE_TEXT_PADDING = 10;
 export const MAX_TREE_PARSING_DEPTH = 300;
 
 export const LABEL_FONT_SIZE = 14;
-
-export const PERMISSIVENESS_SCORE_WEIGHTS = {
-  [LicenseCategory.PERMISSIVE]: 100,
-  [LicenseCategory.WEAK_COPYLEFT]: 60,
-  [LicenseCategory.STRONG_COPYLEFT]: 20,
-  [LicenseCategory.UNKNOWN]: 10,
-};
 
 export const ROOT_PACKAGE_KEY = 'root@0.0.0';
