--- packages/license-kit/visualizer/src/types/LicenseAnalysisResult.ts
@@ -1,6 +1,0 @@
-import { LicenseCategory } from './LicenseCategory';
-import { LicenseStats } from './LicenseStats';
-
-export interface LicenseAnalysisResult extends LicenseStats {
-  categorizedLicenses: Record<string, LicenseCategory>;
-}
