+++ packages/licenses-api/src/index.ts
@@ -1,3 +1,4 @@
-export * from './common';
 export * as Types from './types';
 export * from './utils';
+export * from './permissiveness';
+export * from './constants';
