+++ packages/licenses-api/src/node/utils/__tests__/miscUtils.test.ts
