+++ packages/licenses-api/src/node/utils/__tests__/packageUtils.test.ts
