+++ packages/licenses-api/src/node/utils/index.ts
@@ -1,0 +1,2 @@
+export * as PackageUtils from './packageUtils';
+export * as MiscUtils from './miscUtils';
