+++ packages/licenses-api/src/node/utils/miscUtils.ts
