+++ packages/licenses-api/src/node/utils/packageUtils.ts
@@ -1,10 +1,10 @@
 import fs from 'fs';
 import path from 'path';
 
-import type { License, ScanPackageOptionsFactory } from '../types';
+import type { License, ScanPackageOptionsFactory } from '../../types';
+import { normalizeRepositoryUrl } from '../../utils/repositoryUtils';
 
 import { sha512 } from './miscUtils';
-import { normalizeRepositoryUrl } from './repositoryUtils';
 
 export function getPackageJsonPath(dependency: string, root?: string) {
   const rootsToSearch = [
