+++ packages/licenses-api/src/permissiveness/LicenseCategory.ts
