+++ packages/licenses-api/src/permissiveness/index.ts
@@ -1,0 +1,4 @@
+export * from './constants';
+export * from './LicenseCategory';
+export * from './descriptions';
+export * from './licenseAnalysis';
