+++ packages/licenses-api/src/types/LicenseAnalysisResult.ts
@@ -1,0 +1,7 @@
+import type { LicenseCategory } from '../permissiveness/LicenseCategory';
+
+import type { LicenseStats } from './LicenseStats';
+
+export interface LicenseAnalysisResult extends LicenseStats {
+  categorizedLicenses: Record<string, LicenseCategory>;
+}
