+++ packages/licenses-api/src/types/index.ts
@@ -8,3 +8,5 @@ export * from './ScanPackageOptions';
 export * from './ScanPackageCallContext';
 export * from './DependencyType';
 export * from './ParentPackageInfo';
+export * from './LicenseAnalysisResult';
+export * from './LicenseStats';
