+++ packages/licenses-api/src/utils/index.ts
@@ -1,4 +1,2 @@
-export * as PackageUtils from './packageUtils';
-export * as MiscUtils from './miscUtils';
 export * as RepositoryUtils from './repositoryUtils';
 export * as YamlUtils from './yamlUtils';
